# resources/api-spec-v4.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
  - description: Postman mock server
    url: https://a00459c1-e91f-47ea-9c25-bd9e5268f373.mock.pstmn.io
info:
  description: API specification for the SENG302 2021 project
  version: 4.0.1
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  '/businesses':
    post:
      tags:
        - 'U5: Creating Business Accounts'
        - 'Business Accounts'
      summary: Create a new business entity
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Business'
        description: New business account data, id will be generated automatically by the server
      responses:
        '201':
          description: Business account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  businessId:
                    type: integer
                    description: The ID of the business that has just been created
                    example: 100
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  '/businesses/{id}':
    get:
      tags:
        - 'U5: Creating Business Accounts'
        - 'Business Accounts'
      summary: Retrieve a specific business account
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Business retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/businesses/{id}/makeAdministrator':
    put:
      tags:
        - 'U5: Creating Business Accounts'
        - 'Business Accounts'
      summary: Add an individual as an administrator for a business
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  description: The ID of the user to add as a business administrator
                  example: 100
      responses:
        '200':
          description: Individual added as an administrator successfully
        '400':
          description: The individual could not be added as an administrator, perhaps because it does not exist or is already an administrator.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: "Forbidden: Returned when the user making the request is not the primary administrator of the business AND is not a global application admin."
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/businesses/{id}/removeAdministrator':
    put:
      tags:
        - 'U5: Creating Business Accounts'
        - 'Business Accounts'
      summary: Remove an individual from administering a business
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  description: The ID of the user to remove as a business administrator
                  example: 100
      responses:
        '200':
          description: Inidividual removed from administering successfully
        '400':
          description: The individual could not be removed from administering the business, perhaps because it does not exist, was not an administrator of the business, or is the primary administrator of the business.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: "Forbidden: Returned when the user making the request is not the primary administrator of the business AND is not a global application admin."
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/businesses/{id}/products':
    get:
      tags:
        - 'U15: Product Catalogue'
        - 'Business Accounts'
        - 'Catalogue / Products'
      summary: "Retrieve all the products in a business's catalogue"
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Catalogue retrieved successfully. This could be an empty array.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'The account performing the request is neither an administrator of the business, nor a global application admin.'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    post:
      tags:
        - 'U15: Product Catalogue'
        - 'Business Accounts'
        - 'Catalogue / Products'
      summary: "Add a new product to a business's catalogue"
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  productId:
                    type: string
                    description: The ID of the product that has just been created
                    example: "WATT-420-BEANS"
        '400':
          description: There was some error with the data supplied by the user, appropriate error message(s) should be shown to user
        '403':
          description: 'Forbidden: Returned when a user tries to add a product to business they do not administer AND the user is not a global application admin'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
...
  '/businesses/{businessId}/products/{productId}':
    put:
      tags:
        - 'U17: Modify catalogue entries'
        - 'Catalogue / Products'
      summary: Modify an existing catalogue item
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: businessId
          in: path
          description: ID of business to search catalogue of
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          description: ID (Product Code) of product to try to find in catalogue
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product updated successfully
        '400':
          description: There was some error with the data supplied by the user, appropriate error message(s) should be shown to user
        '403':
          description: 'Forbidden: Returned when a user tries to update a product for a business they do not administer AND the user is not a global application admin'
  '/businesses/{id}/inventory/':
    get:
      tags:
        - 'U19: Create inventory'
        - 'Business Accounts'
        - 'Inventory'
      summary: "Retrieve a business's product inventory"
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Inventory retrieved successfully. This could be an empty array.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItem'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'The account performing the request is neither an administrator of the business, nor a global application admin.'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    post:
      tags:
        - 'U19: Create inventory'
        - 'Business Accounts'
        - 'Inventory'
      summary: "Add a new inventory item to a business's inventory"
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Inventory item created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  inventoryItemId:
                    type: integer
                    description: The ID of the inventory item that has just been created
                    example: 101
        '400':
          description: There was some error with the data supplied by the user, appropriate error message(s) should be shown to user
        '403':
          description: 'Forbidden: Returned when a user tries to add an inventory item to the inventory of a business they do not administer AND the user is not a global application admin'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem'
  '/businesses/{businessId}/inventory/{inventoryItemId}':
    put:
      tags:
        - 'U20: Modify inventory entries'
        - 'Business Accounts'
        - 'Inventory'
      summary: Modify an existing inventory item
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: businessId
          in: path
          description: ID of business to search catalogue of
          required: true
          schema:
            type: integer
        - name: inventoryItemId
          in: path
          description: ID of inventory item to update
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem'
      responses:
        '200':
          description: Inventory item updated successfully
        '400':
          description: There was some error with the data supplied by the user, appropriate error message(s) should be shown to user
        '403':
          description: 'Forbidden: Returned when a user tries to update an inventory item for a business they do not administer AND the user is not a global application admin'
...
  '/businesses/{businessId}/listings/{listingId}':
    patch:
      tags:
        - 'U31: Purchases'
        - 'Listings / sales'
      summary: "Purchase a specific sale listing from a business"
      parameters:
        - name: businessId
          in: path
          description: ID of the business to purchase listing from
          required: true
          schema:
            type: integer
            example: 100
        - name: listingId
          in: path
          description: ID of the specific sale listing to purchase
          required: true
          schema:
            type: integer
            example: 36
      security:
        - CookieAuth:
            - user
      responses:
        '200':
          description: Sale Listing succesfully purchased from business
          content:
            application/json:
              schema:
                type: object
                properties:
                  saleItemId:
                    type: integer
                    description: The ID of the sale listing that has just been purchased
                    example: 100
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/businesses/{businessId}/listings/{listingId}/like':
    patch:
      tags:
        - 'U30: Individual Full Sale Listing'
        - 'Listings / sales'
      summary: "Like a sale listing from a business"
      parameters:
        - name: businessId
          in: path
          description: ID of the business to like the listing from
          required: true
          schema:
            type: integer
            example: 100
        - name: listingId
          in: path
          description: ID of the specific sale listing to purchase
          required: true
          schema:
            type: integer
            example: 36
      security:
        - CookieAuth:
            - user
      responses:
        '200':
          description: Sale Listing succesfully liked
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequestError:
      description: There was some error with the data supplied by the user, appropriate error message(s) should be shown to user
    UnauthorizedError:
      description: Access token is missing or invalid
...
    NotAcceptableError:
      description: 'The requested route does exist (so not a 404) but some part of the request is not acceptable, for example trying to access a resource by an ID that does not exist.'
...
  schemas:
...
    Business:
      type: object
      description: Business account entity
      required:
        - name
        - address
        - businessType
      properties:
        id:
          type: integer
          example: 100
          readOnly: true
          description: Automatically generated and assigned by the server
        administrators:
          type: array
          items:
            $ref: '#/components/schemas/User'
          readOnly: true
        primaryAdministratorId:
          type: integer
          example: 20
          description: "The ID of the 'primary' administrator of the business. This is automatically given to the creator of the business."
        name:
          type: string
          example: 'Lumbridge General Store'
        description:
          type: string
          example: 'A one-stop shop for all your adventuring needs'
        address:
          $ref: '#/components/schemas/Address'
        businessType:
          type: string
          enum:
            - Accommodation and Food Services
            - Retail Trade
            - Charitable organisation
            - Non-profit organisation
        created:
          type: string
          format: date-time
          example: '2020-07-14T14:52:00Z'
          readOnly: true
          description: Date and time when the account was created
...